# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stoclio API
  version: 1.0.0
  description: Market intelligence API. Market data only; no order or trading operations.
servers:
  - url: https://stoclio-api.onrender.com
  - url: http://localhost:5000
paths:
  /health:
    get:
      summary: Service health
      responses:
        '200': { description: Healthy }
        '503': { description: Database unavailable }
  /api/auth/register:
    post:
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthPayload' }
      responses:
        '201': { description: Registered }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthPayload' }
      responses:
        '200': { description: Authenticated }
        '401': { description: Invalid credentials }
  /api/auth/me:
    get:
      summary: Current user
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: User profile }
  /api/market/search:
    get:
      summary: Search instruments across NSE, BSE, NFO and MCX
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 2 } }
      responses:
        '200': { description: Search results }
  /api/market/quote:
    get:
      summary: Fetch market quotes
      parameters:
        - { name: symbols, in: query, schema: { type: string } }
        - { name: mode, in: query, schema: { type: string, enum: [LTP, QUOTE, FULL] } }
      responses:
        '200': { description: Quotes }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    AuthPayload:
      type: object
      required: [email, password]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, format: password }
  responses:
    BadRequest:
      description: Invalid request